Inclusion du connecteur mqtt dans l'API directement + modification pour utiliser sqlite_connector.js comme module
(Je sais que ceci est un gros commit, mais long story short, j'ai eu du trouble à merger mes trucs, j'ai dû reset mon repo parce que je l'ai brisé, j'ai perdu tous mes petits commits :( )

- Ajout d'un .gitignore pour jardin.db (on a pas besoin de ça dans le repo puisque c'est généré par le connecteur sqlite si la db n'est pas existante)
- Renommé le fichier connector.js (le connecteur mqtt) en mqtt_connector.js (pour éviter la confusion avec sqlite_connector.js)
- Modifié le fichier mqtt_connector.js pour qu'il soit utilisé en tant que module (dorénavant, on a juste besoin de faire `node api.js` pour avoir l'API complet).
- Modifié le fichier sqlite_connector.js pour éviter d'avoir des trucs non essentiels dans module.exports (j'ai découvert qu'une fonction dans module.exports peut accéder aux variables qui sont à l'extérieur de module.exports)
