+++ serveur-web/api/.gitignore
@@ -1,0 +1,2 @@
+jardin.db
+node_modules/
