+++ serveur-web/api/sqlite_connector.js
@@ -1,16 +1,13 @@
 var out = module.exports;
 
-//Fonction pour initialiser la BD
-out.dbInit = function(dbname) {
-  out.sqlite3 = require('sqlite3').verbose();
-  out.fs = require('fs');
-  out.db = new out.sqlite3.Database('jardin.db');
-  out.db.serialize(); // --> Pour empêcher les requêtes en parallèle
-  var sql = out.fs.readFileSync('sqlite_dbgen.sql', 'utf8');
-  out.db.exec(sql, function(err){
-    if (err) throw err;
-  });
-}
+var sqlite3 = require('sqlite3').verbose();
+var fs = require('fs');
+var db = new sqlite3.Database('jardin.db');
+db.serialize(); // --> Pour empêcher les requêtes en parallèle
+var sql = fs.readFileSync('sqlite_dbgen.sql', 'utf8');
+db.exec(sql, function(err){
+  if (err) throw err;
+});
 
 //Fonction pour ajouter un bucket dans la base de données
 out.createBucket = function(_name, _ip, cb) {
@@ -26,7 +23,7 @@ out.createBucket = function(_name, _ip, cb) {
   }
 
   //Insertion
-  var requete = out.db.prepare("INSERT INTO buckets (name, ip) VALUES (?, ?)");
+  var requete = db.prepare("INSERT INTO buckets (name, ip) VALUES (?, ?)");
   var id;
   requete.run(_name, _ip, function(){
     cb(this.lastID);
@@ -44,7 +41,7 @@ out.createSensor = function(_name, _type, _bucketId, cb) {
     return;
   }
 
-  var requete = out.db.prepare("INSERT INTO sensors (bucket_id, name, type) VALUES (?, ?, ?)");
+  var requete = db.prepare("INSERT INTO sensors (bucket_id, name, type) VALUES (?, ?, ?)");
   var id;
   requete.run(_bucketId, _name, _type, function(){
     cb(this.lastID);
@@ -62,7 +59,7 @@ out.createValue = function(_value, _sensorId) {
     return;
   }
 
-  var requete = out.db.prepare('INSERT INTO "jardin"."values" (sensor_id, value, timestamp) VALUES (?, ?, ?)');
+  var requete = db.prepare('INSERT INTO "jardin"."values" (sensor_id, value, timestamp) VALUES (?, ?, ?)');
   requete.run(_sensorId, _value, Date.now());
 }
 
@@ -77,7 +74,7 @@ out.createValueWithSensorName = function(_value, _sensorName, _bucketName) {
     return;
   }
   //On va chercher le ID du sensor à partir des noms.
-  var requete = out.db.all('SELECT sensors.id FROM sensors, buckets WHERE sensors.bucket_id = buckets.id AND buckets.name = "' + _bucketName + '" AND sensors.name = "' + _sensorName + '"', function(err, rows){
+  var requete = db.all('SELECT sensors.id FROM sensors, buckets WHERE sensors.bucket_id = buckets.id AND buckets.name = "' + _bucketName + '" AND sensors.name = "' + _sensorName + '"', function(err, rows){
     if (err) throw err;
     out.createValue(_value, rows[0].id);
   });
@@ -87,19 +84,19 @@ out.createValueWithSensorName = function(_value, _sensorName, _bucketName) {
 //-------Fonctions pour aller chercher les informations dans la BD-------
 
 out.getBucketList = function(cb){
-  out.db.all("SELECT * FROM buckets", function(err, rows){
+  db.all("SELECT * FROM buckets", function(err, rows){
     cb(rows);
   });
 }
 
 out.getBucketInfo = function(id, cb){
-  out.db.all("SELECT * FROM sensors WHERE bucket_id = " + id, function(err, rows){
+  db.all("SELECT * FROM sensors WHERE bucket_id = " + id, function(err, rows){
     cb(rows);
   });
 }
 
 out.getSensorValue = function(id, cb){
-  out.db.all('SELECT * FROM "values" WHERE sensor_id = ' + id, function(err, rows){
+  db.all('SELECT * FROM "values" WHERE sensor_id = ' + id, function(err, rows){
     cb(rows);
   });
 }
